requirements.txt
google-adk[a2a]==1.19.0
litellm==1.80.7
openai==2.8.1
langchain-community>=0.4.1